+++ src/components/textfield/TextField.spec.js
@@ -2,6 +2,7 @@ import assert from 'power-assert';
 
 import Harness from '../../../test/harness';
 import TextFieldComponent from './TextField';
+import EventEmitter from '../../EventEmitter';
 
 import {
   comp1
@@ -19,8 +20,22 @@ describe('TextField Component', () => {
 describe('TextField Builder', () => {
   let builder = null;
 
-  before((done) => Harness.builderBefore(done));
-  after(() => Harness.builderAfter());
+  before((done) => {
+    // Incrise Events limit for this tests set
+    Harness.builderBefore(done, {
+      editForm: {
+        events: new EventEmitter({
+          wildcard: false,
+          maxListeners: 0,
+          loadLimit: 250,
+          log: true,
+        })
+      }
+    });
+  });
+  after(() => {
+    Harness.builderAfter();
+  });
 
   it('Should create a new textfield component', (done) => {
     builder = Harness.buildComponent('textfield');