+++ Changelog.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Fixed
  - FOR-2000: POST to URL button not passing headers, added interpolation to header value
+ - FOR-1806: ```initialized``` event firing before first ```change``` event when setting language
+
+### Added
+ - reCAPTCHA component
 
 ## 3.12.3
 ### Fixed
  - FOR-1802: ```change```  event not firing when component position is changed in PDF Form in builder
- - Issue with W``izard cancel.
+ - Issue with Wizard cancel.
  - Issue where an error of dataValue.map is not a function would trigger in file components.
  - Problem with Headers not being defined for non-brower implementations.